reactor: Move make_pollfn<> to internal/poller.hh
